Added sets to domain, query fot JpaRepositories.

+++ semestral_work/src/main/java/cz/cvut/fit/tjv/soukuj26/semestral_work/domain/FitnessCenter.java
@@ -15,7 +15,7 @@ public class FitnessCenter implements Serializable { //Serializable may be used 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "id_fitness_center")
-    private Integer idFitnessCenter;
+    private Long idFitnessCenter;
 
     @Column(name = "name", nullable = false)
     private String name;
